# planung/restapi/api/src/main/java/org/openapitools/api/WaveSensorApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.1.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.PresenceLog;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2023-11-19T14:12:38.522563+01:00[Europe/Berlin]")
@Validated
@Tag(name = "WaveSensor", description = "the WaveSensor API")
public interface WaveSensorApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /wecker/sensor/presence/logs : Add Sensor Presence Log entry
     * Add Sensor Presence Log entry
     *
     * @param presenceLog Sensor Presence Log Entry (optional)
     * @return Successful operation (status code 200)
     *         or Invalid ID supplied (status code 400)
     *         or Sensor Presence Log not found (status code 404)
     *         or Validation exception (status code 405)
     */
    @Operation(
        operationId = "addPresenceLog",
        summary = "Add Sensor Presence Log entry",
        description = "Add Sensor Presence Log entry",
        tags = { "WaveSensor" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful operation", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PresenceLog.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid ID supplied"),
            @ApiResponse(responseCode = "404", description = "Sensor Presence Log not found"),
            @ApiResponse(responseCode = "405", description = "Validation exception")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/wecker/sensor/presence/logs",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<PresenceLog> addPresenceLog(
        @Parameter(name = "PresenceLog", description = "Sensor Presence Log Entry") @Valid @RequestBody(required = false) PresenceLog presenceLog
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"areaId\" : 10, \"statusId\" : 10, \"id\" : 10, \"timestamp\" : 10 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /wecker/sensor/presence/logs/{logId} : Get Sensor Presence Log by ID
     * Get Sensor Presence Log by ID
     *
     * @param logId ID of Log (required)
     * @return Successful operation (status code 200)
     *         or Invalid ID supplied (status code 400)
     *         or Sensor Presence Log not found (status code 404)
     *         or Validation exception (status code 405)
     */
    @Operation(
        operationId = "getPresenceLogById",
        summary = "Get Sensor Presence Log by ID",
        description = "Get Sensor Presence Log by ID",
        tags = { "WaveSensor" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful operation", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PresenceLog.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid ID supplied"),
            @ApiResponse(responseCode = "404", description = "Sensor Presence Log not found"),
            @ApiResponse(responseCode = "405", description = "Validation exception")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/wecker/sensor/presence/logs/{logId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<PresenceLog> getPresenceLogById(
        @Parameter(name = "logId", description = "ID of Log", required = true, in = ParameterIn.PATH) @PathVariable("logId") Long logId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"areaId\" : 10, \"statusId\" : 10, \"id\" : 10, \"timestamp\" : 10 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /wecker/sensor/presence/logs : Get all Sensor presence Logs
     * Get all Sensor presence Logs
     *
     * @return Successful operation (status code 200)
     */
    @Operation(
        operationId = "getPresenceLogs",
        summary = "Get all Sensor presence Logs",
        description = "Get all Sensor presence Logs",
        tags = { "WaveSensor" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful operation", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = PresenceLog.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/wecker/sensor/presence/logs",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<PresenceLog>> getPresenceLogs(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"areaId\" : 10, \"statusId\" : 10, \"id\" : 10, \"timestamp\" : 10 }, { \"areaId\" : 10, \"statusId\" : 10, \"id\" : 10, \"timestamp\" : 10 } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}