# planung/restapi/api/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  description: Dies ist eine Schnittstellenbeschreibung für die Wecker-REST-API
  termsOfService: http://swagger.io/terms/
  title: Wecker Rest API - OpenAPI 3.0
  version: 3.1.0
servers:
- url: localhost/api/v1
...
paths:
...
  /wecker/sensor/presence/logs:
    get:
      description: Get all Sensor presence Logs
      operationId: getPresenceLogs
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/PresenceLog'
                type: array
          description: Successful operation
      summary: Get all Sensor presence Logs
      tags:
      - WaveSensor
      x-accepts: application/json
      x-tags:
      - tag: WaveSensor
    post:
      description: Add Sensor Presence Log entry
      operationId: addPresenceLog
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PresenceLog'
        description: Sensor Presence Log Entry
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresenceLog'
          description: Successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Sensor Presence Log not found
        "405":
          description: Validation exception
      summary: Add Sensor Presence Log entry
      tags:
      - WaveSensor
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: WaveSensor
  /wecker/sensor/presence/logs/{logId}:
    get:
      description: Get Sensor Presence Log by ID
      operationId: getPresenceLogById
      parameters:
      - description: ID of Log
        explode: false
        in: path
        name: logId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresenceLog'
          description: Successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Sensor Presence Log not found
        "405":
          description: Validation exception
      summary: Get Sensor Presence Log by ID
      tags:
      - WaveSensor
      x-accepts: application/json
      x-tags:
      - tag: WaveSensor
...
  /wecker/stati/:
    get:
      description: Get all Statis
      operationId: getAllStatis
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Status'
                type: array
          description: Successful operation
      summary: Get all Statis
      tags:
      - Status
      x-accepts: application/json
      x-tags:
      - tag: Status
    post:
      description: Add Status with ID
      operationId: addStatus
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Status'
        description: Add Status
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
          description: Successful operation
        "400":
          description: Invalid ID supplied
        "405":
          description: Validation exception
      summary: Add Status with ID
      tags:
      - Status
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Status
  /wecker/stati/{statusId}:
    get:
      description: Get Status via ID
      operationId: getStatusViaId
      parameters:
      - description: ID of Status
        explode: false
        in: path
        name: statusId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
          description: Successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Status not found
        "405":
          description: Validation exception
      summary: Get Status via ID
      tags:
      - Status
      x-accepts: application/json
      x-tags:
      - tag: Status
    put:
      description: Update Status with ID
      operationId: updateStatusViaId
      parameters:
      - description: ID of Status
        explode: false
        in: path
        name: statusId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Status'
        description: Update an existent Status
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
          description: Successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Status not found
        "405":
          description: Validation exception
      summary: Update Status with ID
      tags:
      - Status
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Status
components:
...
  schemas:
...
    Status:
      example:
        name: Mensch Anwesend
        id: 123
      properties:
        id:
          example: 123
          format: int64
          type: integer
        name:
          example: Mensch Anwesend
          type: string
      type: object
...
    PresenceLog:
      example:
        areaId: 10
        statusId: 10
        id: 10
        timestamp: 10
      properties:
        id:
          example: 10
          format: int64
          type: integer
        areaId:
          example: 10
          format: int64
          type: integer
        statusId:
          example: 10
          format: int64
          type: integer
        timestamp:
          example: 10
          format: int64
          type: integer
      type: object